refactor(context-xray): drop the formatter re-export and redundant test rows
format.ts re-exported formatTokenCount purely so the three x-ray components
could import it and formatPercentOf from one place. The indirection bought
nothing and needed a test asserting it was still the same function object;
the components now import the canonical formatter directly and both the
re-export and its guard test are gone.

Also trims table rows that exercise a branch already covered: three
percent cases that all hit the total <= 0 guard or the sub-1% floor, two
precise-mode values above the single 999_500 rollover boundary, one
out-of-range threshold variant, and a concurrency test that repeated the
threshold scenario on the cutoff field without crossing fields, which is
the only part the per-field write guard adds.
